refactor: add rollback support to chain_store and organize layout
